+++ TombEngine/Game/control/control.cpp
@@ -669,11 +669,8 @@ GameStatus HandleMenuCalls(bool isTitle)
 	else if (doLoad && g_GameFlow->IsLoadSaveEnabled() && g_Gui.GetInventoryMode() != InventoryMode::Load)
 	{
 		SaveGame::LoadHeaders();
-
 		g_Gui.SetInventoryMode(InventoryMode::Load);
-
-		if (g_Gui.CallInventory(LaraItem, false))
-			gameStatus = GameStatus::LoadGame;
+		g_Gui.CallInventory(LaraItem, false);
 	}
 	else if (doPause && g_Gui.GetInventoryMode() != InventoryMode::Pause)
 	{